Currently, the pantry app is tightly coupled with the main waitress app. This commit creates a pantry app and moves the pantry viewset into the pantry app
