+++ waitress/pantry/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class PantryConfig(AppConfig):
+    name = 'pantry'
